Introduce API key-style secret support to server
An optional `--api-secret`/`store.api_seret` option on the client side.
If provided, the client includes it with all request.

An optional `TIR_API_SECRET` environment variable on the server side. If
not provided, the server doesn't authenticate requests. If provided, the
server requires it be set for all non-GET requests (create, update,
delete).

Uses a bearer scheme:

```
Authentication: Bearer YOUR_API_SECRET
```