Add draft MySQL store, tests
The prepared SQL statements here are MySQL specific, and the Golang
MySQL driver I'm using (though seemingly fairly official) leaves a lot
to be desired.

Benchmark needs to be modified: each time we run the benchmark, we need
to generate a *fresh* store (out of timer) so that the times will
converge.