REVERTME: a bad benchmark
Why bad? Testing an in-memory SQL driver isn't really fair; it's unclear
how performant it is vs. actual MySQL. Moreover, dynamics in practice
are a little different: if a store is being loaded to serve the request,
the whole file has to be parsed; in this test, the file is parsed once
and then overwritten. That might also explain why the file perf doesn't
scale with initial file size: I should close and reopen it.

Anyway, the store benchmark concept was interesting; I learned something
here.